backend/requirements.txt
Flask
flask-cors
pytesseract
pdf2image
requests
python-dotenv
firebase-admin
gunicorn
